+++ src/sql/error_handler.py
@@ -47,7 +47,9 @@ def _detailed_message_with_error_type(error, query):
         "invalid sql",
     ]
     not_found_substrings = [
-        "does not exist",
+        r"(\btable with name\b).+(\bdoes not exist\b)",
+        r"(\brelation\b).+(\bdoes not exist\b)",
+        r"(\btable\b).+(\bdoesn't exist\b)",
         "not found",
         "could not find",
         "no such table",
@@ -63,7 +65,10 @@ def _detailed_message_with_error_type(error, query):
                 TableNotFoundError,
             )
         else:
-            return f"{CTE_MSG}\n\n{ORIGINAL_ERROR}{original_error}\n", RuntimeError
+            return (
+                f"{CTE_MSG}\n\n{ORIGINAL_ERROR}{original_error}\n",
+                RuntimeError,
+            )
     elif "fe_sendauth: no password supplied" in original_error:
         return f"{POSTGRES_MSG}\n{ORIGINAL_ERROR}{original_error}\n", RuntimeError
     return None, None